Fix the fzf plugin
Before this change, fzf would fail when I tried using any of the vim
functions. The key is that there are two plugins. The one that is the
submodule enables the helper functions in vim like :Files, but the
second plugin, which is packaged with fzf, is the one that this plugin
depends on. The plugin that is in the fzf application repo needs to be
sourced.