Remove local overrides
Before this idea, I planned to put local overrides in an optional .local
file, but this is unnecessary. This can be better managed with a git
branch. To avoid pushing (for instance on my work laptop), I can turn
off pushing to remote on the branch using
`git remote set-url --push origin no_push`. This lets me pull but if I
try to push, it will push to the url no_push and fail.