Use s for search instead of substitute
Prior to this change, s was used for substitute in normal mode, but this
functionality is easily duplicated with c, and I don't use it
frequently. Search is more useful in this context. This change noops s
on its own and adds a number of search options that depend on fzf.

Closes https://github.com/mbmjensen/vim/issues/3