Separate which key into visual and normal mappings
Before this, the same mappings were used for normal and visual mode.
With this change, the hints are different for both modes. It is worth
considering doing the same thing for other mappings that are easy to
forget such as 'g'.

This change also included some small code cleanups and reformatting,
such as making all key codes lower case and using <space> literally
instead of using the leader key.