Update Coc popup menu for suggestions
Do not select anything when the popup menu appears. It triggers
automatically which makes it difficult to select the first item with the
default setting. With this setup, you can see the popup menu
automatically but you must press C-n to select the first result.

This change also adds virtualtext for lsps that support it. If this gets
annoying it can be turned off later.