Select first suggestion by default
This change makes it so that coc selects the first completion suggestion
by default. To accept this selection, press Tab. The change also removes
the virtual text for this suggestion since my colorscheme does not
distinguish between the appearance of the real text and the virtual text
which made the hint confusing.